api call moved to tabs container and Facets tab refactoring
External component functions in FacetsTab moved to specific ui components to avoid unnecessary rerendering.

Implementation of resultsCount in General data using api data and skeleton implementation
