+++ src/components/tabs/FacetsTab.tsx
@@ -1,62 +1,24 @@
-import { Box, List, TabPanel, SkeletonCircle } from "@chakra-ui/react";
+import { Box, TabPanel} from "@chakra-ui/react";
 import { PieChart } from "./charts/PieChart";
-import { getApi } from "../../api/api";
-import { useState, useEffect } from "react";
-import { ApiResponse } from "../../types/api";
-import { ListItemComponent } from "../ListItemComponent";
-import { ListItemSkeleton } from "../ListItemSkeleton";
 import { PieLoading } from "./ui/PieLoading";
+import { Format } from "../../types/api";
+import { ListFormats } from "./ui/list/ListFormats";
 
-export const FacetsTab = ({ format }) => {
-  const [data, setData] = useState<ApiResponse>();
+type Props = {
+  format: Format[] | undefined;
+}
 
-  useEffect(() => {
-    const fetchData = async () => {
-      try {
-        const resp = await getApi();
-        setData(resp);
-        console.log(resp);
-      } catch (error) {
-        console.error(error);
-      }
-    };
-    fetchData();
-  }, []);
-
-  //get format from data
-
-  const getFormatFromData = (data: ApiResponse) => data.facets.format;
-
-
-  const renderList = () => {
-    return data
-      ? getFormatFromData(data).map((item) => (
-          <ListItemComponent item={item} key={item.translated} />
-        ))
-      : renderListSkeleton();
-  };
-
-  //List skeleton
-  const renderListSkeleton = () => {
-    return (
-      //iterate 8 times to simulate a list
-      Array.from({ length: 8 }).map(() => <ListItemSkeleton />)
-    );
-  };
+export const FacetsTab = ({ format }:Props) => {
 
   
   return (
     <TabPanel p="0" display="flex">
-      <Box bgColor="#7FC7BD" w="65%" p="6" h="auto" display="flex" alignItems="center"
-      >
-        <List listStyleType="-" display="flex" flexWrap="wrap" gap="4" flexDirection="row" p={2}
-        >
-          { /* { <ListLoading  /> } */}
-        </List>
+      <Box bgColor="#7FC7BD" w="65%" p="6" h="auto" display="flex" alignItems="center">
+          <ListFormats format={format}  /> 
       </Box>
       <Box bgColor="#b4e2dc" w="35%" h="270px">
-        {data ? (
-          <PieChart data={getFormatFromData(data)} />
+        {format ? (
+          <PieChart data={format} />
         ) : (
           <PieLoading />
         )}
