+++ src/components/tabs/GeneralData.tsx
@@ -1,12 +1,10 @@
-import { Box, Divider, Text } from '@chakra-ui/react'
+import { Box, Divider, Text, SkeletonText } from '@chakra-ui/react'
 
 type Props = {
-  results: number;
+  results: number | undefined;
 }
 
 export const GeneralData = ({ results }:Props ) => {
-
-  
   return (
     <Box bgColor="#68b9b7" w="33.33%" py="10" borderTopLeftRadius={8} borderBottomLeftRadius={8} display="flex" justifyContent="center" alignItems="center" flexDir="column" >
       <Box>
@@ -16,10 +14,13 @@ export const GeneralData = ({ results }:Props ) => {
       </Box>
       <Divider my="3" w="50%" />
       <Box>
-        <Text as="span" fontSize="3xl" fontWeight="semibold" color="white">
-          <b>5.016.521</b> - Documentos 
-          {/* Show resultsCounts */}
-        </Text>
+        {results ? (
+          <Text as="span" fontSize="3xl" fontWeight="semibold" color="white">
+            <b>{results.toLocaleString()}</b> - Documentos 
+          </Text>
+        ) : (
+          <SkeletonText noOfLines={1} spacing="4" width="20em" />
+        )}
       </Box>
     </Box>
   )
