--- src/components/tabs/ui/ListLoading.tsx
@@ -1,6 +1,0 @@
-
-export const ListLoading = () => {
-  return (
-    <div>ListLoading</div>
-  )
-}
